chores: move yCRV out for consistency
