+++ apps/common/components/AppHeader.tsx
@@ -12,7 +12,6 @@ import {cl} from '@yearn-finance/web-lib/utils/cl';
 import {useMenu} from '@common/contexts/useMenu';
 import {useCurrentApp} from '@common/hooks/useCurrentApp';
 import {LogoYearn} from '@common/icons/LogoYearn';
-import {YCrvHeader} from '@yCRV/components/header/YCrvHeader';
 
 import {AppName, APPS} from './Apps';
 import {MotionDiv} from './MotionDiv';
@@ -25,7 +24,6 @@ function Logo(): ReactElement {
 
 	return (
 		<>
-			<YCrvHeader pathname={pathname} />
 			<VaultsHeader pathname={pathname} />
 			<VeYfiHeader pathname={pathname} />
 			<VaultsV3Header pathname={pathname} />
@@ -159,10 +157,6 @@ export function AppHeader(): ReactElement {
 	const {onOpenMenu} = useMenu();
 	const menu = useMemo((): TMenu[] => {
 		const HOME_MENU = {path: '/', label: 'Home'};
-
-		if (pathname.startsWith('/ycrv')) {
-			return [HOME_MENU, ...APPS[AppName.YCRV].menu];
-		}
 
 		if (pathname.startsWith('/v3')) {
 			return [HOME_MENU, ...APPS[AppName.VAULTSV3].menu];
