+++ apps/common/hooks/useCurrentApp.tsx
@@ -3,7 +3,6 @@ import {VeYfiHeader} from 'apps/veyfi/components/header/VeYfiHeader';
 import homeManifest from 'public/manifest.json';
 import {VaultsHeader} from '@vaults/components/header/VaultsHeader';
 import {AppName, APPS} from '@common/components/Apps';
-import {YCrvHeader} from '@yCRV/components/header/YCrvHeader';
 
 import type {NextRouter} from 'next/router';
 import type {ReactElement} from 'react';
@@ -28,10 +27,6 @@ export function useCurrentApp({pathname}: NextRouter): TCurrentApp {
 			'/vaults': {
 				...APPS[AppName.VAULTS],
 				header: <VaultsHeader pathname={pathname} />
-			},
-			'/ycrv': {
-				...APPS[AppName.YCRV],
-				header: <YCrvHeader pathname={pathname} />
 			},
 			'/veyfi': {
 				...APPS[AppName.VEYFI],
