+++ apps/common/utils/index.tsx
@@ -1,61 +1,5 @@
-import {request} from 'graphql-request';
-import {formatUnits, parseUnits} from 'viem';
-import {captureException} from '@sentry/nextjs';
-import {toAddress} from '@yearn-finance/web-lib/utils/address';
-import {
-	LPYCRV_TOKEN_ADDRESS,
-	LPYCRV_V2_TOKEN_ADDRESS,
-	YCRV_CURVE_POOL_ADDRESS,
-	YVBOOST_TOKEN_ADDRESS,
-	YVECRV_TOKEN_ADDRESS
-} from '@yearn-finance/web-lib/utils/constants';
-import {formatToNormalizedValue, toBigInt} from '@yearn-finance/web-lib/utils/format.bigNumber';
-import {formatPercent} from '@yearn-finance/web-lib/utils/format.number';
-
-import type {GraphQLResponse} from 'graphql-request/build/esm/types';
 import type {TDict} from '@yearn-finance/web-lib/types';
 import type {TYDaemonVault} from '@common/schemas/yDaemonVaultsSchemas';
-
-export function max(input: bigint, balance: bigint): bigint {
-	if (input > balance) {
-		return balance;
-	}
-	return input;
-}
-
-export function getVaultAPR(vaults: TDict<TYDaemonVault | undefined>, vaultAddress: string): string {
-	if (!vaults?.[toAddress(vaultAddress)]) {
-		return '';
-	}
-
-	if (toAddress(vaultAddress) === YVECRV_TOKEN_ADDRESS || toAddress(vaultAddress) === YVBOOST_TOKEN_ADDRESS) {
-		return `APY ${formatPercent(0)}`;
-	}
-
-	if (vaults?.[toAddress(vaultAddress)]?.apr?.netAPR) {
-		return `APY ${formatPercent((vaults?.[toAddress(vaultAddress)]?.apr?.netAPR || 0) * 100, 2, 2, 500)}`;
-	}
-	return `APY ${formatPercent(0)}`;
-}
-
-export function getAmountWithSlippage(from: string, to: string, value: bigint, slippage: number): number {
-	const hasLP =
-		toAddress(from) === LPYCRV_TOKEN_ADDRESS ||
-		toAddress(to) === LPYCRV_TOKEN_ADDRESS ||
-		toAddress(from) === LPYCRV_V2_TOKEN_ADDRESS ||
-		toAddress(to) === LPYCRV_V2_TOKEN_ADDRESS;
-	const isDirectDeposit =
-		toAddress(from) === YCRV_CURVE_POOL_ADDRESS ||
-		toAddress(to) === LPYCRV_TOKEN_ADDRESS ||
-		toAddress(to) === LPYCRV_V2_TOKEN_ADDRESS;
-
-	if (hasLP && !isDirectDeposit) {
-		const minAmountStr = Number(formatUnits(toBigInt(value), 18));
-		const minAmountWithSlippage = parseUnits((minAmountStr * (1 - slippage / 100)).toFixed(18) as `${number}`, 18);
-		return formatToNormalizedValue(toBigInt(minAmountWithSlippage), 18);
-	}
-	return formatToNormalizedValue(value, 18);
-}
 
 export function getVaultName(vault: TYDaemonVault): string {
 	const baseName = vault.name;
@@ -64,11 +8,6 @@ export function getVaultName(vault: TYDaemonVault): string {
 	}
 	return baseName;
 }
-
-export const graphFetcher = async (args: [string, string]): Promise<GraphQLResponse> => {
-	const [url, query] = args;
-	return request(url, query);
-};
 
 export function formatDateShort(value: number): string {
 	let locale = 'fr-FR';
@@ -96,13 +35,4 @@ export async function hash(message: string): Promise<string> {
 	const hashArray = Array.from(new Uint8Array(hashBuffer)); // convert buffer to byte array
 	const hashHex = hashArray.map((b): string => b.toString(16).padStart(2, '0')).join(''); // convert bytes to hex string
 	return `0x${hashHex}`;
-}
-
-export function handleSettle<T>(data: PromiseSettledResult<unknown>, fallback: T): T {
-	if (data.status !== 'fulfilled') {
-		console.error(data.reason);
-		captureException(data.reason);
-		return fallback;
-	}
-	return data.value as T;
 }
