--- apps/ycrv/components/illustrations/how-it-works.tsx
@@ -1,189 +1,0 @@
-import type {ReactElement} from 'react';
-
-export function HowItWorksDiagram(props: React.SVGProps<SVGSVGElement>): ReactElement {
-	return (
-		<svg
-			{...props}
-			width={'100%'}
-			height={'100%'}
-			viewBox={'0 0 980 288'}
-			fill={'none'}
-			xmlns={'http://www.w3.org/2000/svg'}>
-			<rect
-				x={'475'}
-				y={'64'}
-				width={'40'}
-				height={'80'}
-				fill={'#9D9D9D'}
-				stroke={'#0C0C0C'}
-			/>
-			<rect
-				x={'115'}
-				y={'144'}
-				width={'200'}
-				height={'80'}
-				fill={'#EBEBEB'}
-				stroke={'#0C0C0C'}
-			/>
-			<rect
-				x={'315'}
-				y={'144'}
-				width={'200'}
-				height={'80'}
-				fill={'#EBEBEB'}
-				stroke={'#0C0C0C'}
-			/>
-			<rect
-				x={'515'}
-				y={'144'}
-				width={'200'}
-				height={'80'}
-				fill={'#EBEBEB'}
-				stroke={'#0C0C0C'}
-			/>
-			<rect
-				x={'715'}
-				y={'144'}
-				width={'200'}
-				height={'80'}
-				fill={'#EBEBEB'}
-				stroke={'#0C0C0C'}
-			/>
-			<rect
-				x={'115'}
-				y={'64'}
-				width={'400'}
-				height={'80'}
-				fill={'#E1E1E1'}
-				stroke={'#0C0C0C'}
-			/>
-			<rect
-				x={'475'}
-				y={'64'}
-				width={'40'}
-				height={'80'}
-				fill={'#9D9D9D'}
-				stroke={'#0C0C0C'}
-			/>
-			<rect
-				x={'515'}
-				y={'64'}
-				width={'400'}
-				height={'80'}
-				fill={'#E1E1E1'}
-				stroke={'#0C0C0C'}
-			/>
-			<path
-				d={
-					'M431.836 44V37.52L436.348 44H437.308V35.6H436.288V42.08L431.776 35.6H430.816V44H431.836ZM441.613 44.072C443.401 44.072 444.649 42.8 444.649 40.94C444.649 39.08 443.401 37.808 441.613 37.808C439.825 37.808 438.577 39.08 438.577 40.94C438.577 42.8 439.825 44.072 441.613 44.072ZM441.613 43.196C440.401 43.196 439.585 42.248 439.585 40.94C439.585 39.632 440.401 38.684 441.613 38.684C442.825 38.684 443.641 39.632 443.641 40.94C443.641 42.248 442.825 43.196 441.613 43.196ZM446.982 44H448.062L449.598 39.452L451.074 44H452.154L454.194 37.88H453.198L451.602 42.86L450.078 37.88H449.166L447.558 42.86L446.034 37.88H445.002L446.982 44Z'
-				}
-				fill={'currentcolor'}
-			/>
-			<path
-				d={
-					'M421.049 244.512C421.049 244.932 420.857 245.232 420.461 245.232H419.093V246.132H420.449C420.713 246.132 420.905 246.084 421.025 245.976V252H422.033V243.6H421.049V244.512ZM422.949 244.524H427.773L423.657 252H424.749L428.877 244.308V243.6H422.949V244.524ZM434.486 249.936C434.486 250.692 434.174 251.04 433.418 251.04H432.458V252H433.55C434.846 252 435.53 251.316 435.53 250.02V243.6H434.486V249.936ZM442.426 251.1C442.15 251.1 442.03 251.016 442.03 250.692V248.016C442.03 246.588 441.13 245.808 439.57 245.808C438.094 245.808 437.122 246.516 436.978 247.716H437.962C438.082 247.092 438.658 246.684 439.522 246.684C440.482 246.684 441.046 247.164 441.046 247.956V248.376H439.246C437.638 248.376 436.786 249.06 436.786 250.284C436.786 251.4 437.698 252.072 439.042 252.072C440.098 252.072 440.71 251.604 441.094 251.052C441.106 251.652 441.382 252 442.198 252H442.738V251.1H442.426ZM441.046 249.504C441.046 250.524 440.338 251.232 439.114 251.232C438.298 251.232 437.782 250.824 437.782 250.224C437.782 249.528 438.274 249.204 439.15 249.204H441.046V249.504ZM446.494 245.808C445.462 245.808 444.838 246.24 444.478 246.708L444.346 245.88H443.494V252H444.478V248.94C444.478 247.56 445.174 246.708 446.326 246.708C447.394 246.708 447.982 247.38 447.982 248.652V252H448.966V248.592C448.966 246.564 447.874 245.808 446.494 245.808ZM452.953 252H458.737V251.076H454.345L456.865 248.94C457.861 248.088 458.677 247.296 458.677 246.036C458.677 244.524 457.621 243.528 455.809 243.528C454.117 243.528 453.001 244.656 453.001 246.24V246.456H454.009V246.3C454.009 245.148 454.729 244.428 455.809 244.428C456.925 244.428 457.645 245.028 457.645 246.084C457.645 246.996 456.997 247.644 456.217 248.316L452.953 251.148V252ZM463.899 247.632C464.703 247.416 465.255 246.804 465.255 245.748C465.255 244.428 464.163 243.528 462.435 243.528C460.779 243.528 459.627 244.548 459.615 246.024H460.611C460.623 245.064 461.391 244.428 462.447 244.428C463.455 244.428 464.223 244.956 464.223 245.808C464.223 246.66 463.599 247.284 462.351 247.284H461.691V248.16H462.411C463.719 248.16 464.439 248.724 464.439 249.6C464.439 250.56 463.599 251.172 462.507 251.172C461.319 251.172 460.455 250.512 460.443 249.408H459.447C459.447 251.076 460.623 252.072 462.447 252.072C464.271 252.072 465.471 251.1 465.471 249.672C465.471 248.388 464.667 247.776 463.899 247.632Z'
-				}
-				fill={'currentcolor'}
-			/>
-			<path
-				d={
-					'M453.004 276.512C453.004 276.932 452.812 277.232 452.416 277.232H451.048V278.132H452.404C452.668 278.132 452.86 278.084 452.98 277.976V284H453.988V275.6H453.004V276.512ZM458.216 284.072C460.22 284.072 461.408 282.404 461.408 279.764C461.408 278.66 461.168 277.652 460.748 276.92C460.22 276.056 459.368 275.528 458.204 275.528C456.476 275.528 455.288 276.728 455.288 278.348C455.288 280.064 456.452 281.156 458.108 281.156C459.236 281.156 460.076 280.604 460.448 279.776C460.496 279.668 460.508 279.596 460.52 279.548C460.532 279.752 460.532 279.836 460.532 279.956C460.532 281.852 459.728 283.184 458.264 283.184C457.244 283.184 456.644 282.608 456.452 281.756H455.444C455.648 283.16 456.692 284.072 458.216 284.072ZM458.18 280.268C457.016 280.268 456.296 279.464 456.296 278.348C456.296 277.22 457.052 276.416 458.216 276.416C459.368 276.416 460.136 277.196 460.136 278.324C460.136 279.476 459.344 280.268 458.18 280.268ZM467.403 281.936C467.403 282.692 467.091 283.04 466.335 283.04H465.375V284H466.467C467.763 284 468.447 283.316 468.447 282.02V275.6H467.403V281.936ZM475.342 283.1C475.066 283.1 474.946 283.016 474.946 282.692V280.016C474.946 278.588 474.046 277.808 472.486 277.808C471.01 277.808 470.038 278.516 469.894 279.716H470.878C470.998 279.092 471.574 278.684 472.438 278.684C473.398 278.684 473.962 279.164 473.962 279.956V280.376H472.162C470.554 280.376 469.702 281.06 469.702 282.284C469.702 283.4 470.614 284.072 471.958 284.072C473.014 284.072 473.626 283.604 474.01 283.052C474.022 283.652 474.298 284 475.114 284H475.654V283.1H475.342ZM473.962 281.504C473.962 282.524 473.254 283.232 472.03 283.232C471.214 283.232 470.698 282.824 470.698 282.224C470.698 281.528 471.19 281.204 472.066 281.204H473.962V281.504ZM479.41 277.808C478.378 277.808 477.754 278.24 477.394 278.708L477.262 277.88H476.41V284H477.394V280.94C477.394 279.56 478.09 278.708 479.242 278.708C480.31 278.708 480.898 279.38 480.898 280.652V284H481.882V280.592C481.882 278.564 480.79 277.808 479.41 277.808ZM485.869 284H491.653V283.076H487.261L489.781 280.94C490.777 280.088 491.593 279.296 491.593 278.036C491.593 276.524 490.537 275.528 488.725 275.528C487.033 275.528 485.917 276.656 485.917 278.24V278.456H486.925V278.3C486.925 277.148 487.645 276.428 488.725 276.428C489.841 276.428 490.561 277.028 490.561 278.084C490.561 278.996 489.913 279.644 489.133 280.316L485.869 283.148V284ZM496.815 279.632C497.619 279.416 498.171 278.804 498.171 277.748C498.171 276.428 497.079 275.528 495.351 275.528C493.695 275.528 492.543 276.548 492.531 278.024H493.527C493.539 277.064 494.307 276.428 495.363 276.428C496.371 276.428 497.139 276.956 497.139 277.808C497.139 278.66 496.515 279.284 495.267 279.284H494.607V280.16H495.327C496.635 280.16 497.355 280.724 497.355 281.6C497.355 282.56 496.515 283.172 495.423 283.172C494.235 283.172 493.371 282.512 493.359 281.408H492.363C492.363 283.076 493.539 284.072 495.363 284.072C497.187 284.072 498.387 283.1 498.387 281.672C498.387 280.388 497.583 279.776 496.815 279.632Z'
-				}
-				fill={'currentcolor'}
-			/>
-			<path
-				d={
-					'M887.516 279.992C887.516 282.5 888.812 284.072 890.708 284.072C892.592 284.072 893.888 282.5 893.888 279.992V279.608C893.888 277.22 892.592 275.528 890.708 275.528C888.812 275.528 887.516 277.22 887.516 279.608V279.992ZM892.868 279.956C892.868 281.948 892.004 283.148 890.708 283.148C889.4 283.148 888.536 281.948 888.536 279.956V279.644C888.536 277.652 889.4 276.452 890.708 276.452C892.004 276.452 892.868 277.652 892.868 279.644V279.956ZM894.67 284H900.454V283.076H896.062L898.582 280.94C899.578 280.088 900.394 279.296 900.394 278.036C900.394 276.524 899.338 275.528 897.526 275.528C895.834 275.528 894.718 276.656 894.718 278.24V278.456H895.726V278.3C895.726 277.148 896.446 276.428 897.526 276.428C898.642 276.428 899.362 277.028 899.362 278.084C899.362 278.996 898.714 279.644 897.934 280.316L894.67 283.148V284ZM904.785 284H905.829V280.328H909.285V279.368H905.829V276.56H910.005V275.6H904.785V284ZM913.344 284.072C914.76 284.072 915.768 283.172 916.008 281.948H915.012C914.808 282.74 914.136 283.196 913.272 283.196C912.108 283.196 911.376 282.38 911.34 281.216V281.132H916.092C916.104 280.988 916.116 280.844 916.116 280.712C916.068 278.9 914.94 277.808 913.212 277.808C911.484 277.808 910.308 279.08 910.308 280.94C910.308 282.812 911.508 284.072 913.344 284.072ZM911.4 280.316C911.496 279.32 912.324 278.672 913.224 278.672C914.244 278.672 914.964 279.26 915.06 280.316H911.4ZM920.322 277.808C919.29 277.808 918.63 278.3 918.222 278.864V275.6H917.238V284H918.09L918.222 283.064C918.594 283.652 919.29 284.072 920.322 284.072C922.014 284.072 923.262 282.812 923.262 280.94C923.262 278.948 922.014 277.808 920.322 277.808ZM920.226 283.196C919.026 283.196 918.21 282.26 918.21 280.916C918.21 279.608 919.026 278.684 920.226 278.684C921.438 278.684 922.254 279.608 922.254 280.94C922.254 282.272 921.438 283.196 920.226 283.196ZM927.119 284H932.903V283.076H928.511L931.031 280.94C932.027 280.088 932.843 279.296 932.843 278.036C932.843 276.524 931.787 275.528 929.975 275.528C928.283 275.528 927.167 276.656 927.167 278.24V278.456H928.175V278.3C928.175 277.148 928.895 276.428 929.975 276.428C931.091 276.428 931.811 277.028 931.811 278.084C931.811 278.996 931.163 279.644 930.383 280.316L927.119 283.148V284ZM938.065 279.632C938.869 279.416 939.421 278.804 939.421 277.748C939.421 276.428 938.329 275.528 936.601 275.528C934.945 275.528 933.793 276.548 933.781 278.024H934.777C934.789 277.064 935.557 276.428 936.613 276.428C937.621 276.428 938.389 276.956 938.389 277.808C938.389 278.66 937.765 279.284 936.517 279.284H935.857V280.16H936.577C937.885 280.16 938.605 280.724 938.605 281.6C938.605 282.56 937.765 283.172 936.673 283.172C935.485 283.172 934.621 282.512 934.609 281.408H933.613C933.613 283.076 934.789 284.072 936.613 284.072C938.437 284.072 939.637 283.1 939.637 281.672C939.637 280.388 938.833 279.776 938.065 279.632Z'
-				}
-				fill={'currentcolor'}
-			/>
-			<path
-				d={
-					'M417.352 12H418.54L420.832 5.88H419.8L417.964 10.884L416.116 5.88H415.06L417.352 12ZM421.567 12H422.551V3.6H421.567V12ZM427.847 9V8.172H423.851V9H427.847ZM430.678 12.036L430.378 12.78C430.15 13.356 430.054 13.5 429.49 13.5H428.614V14.4H429.79C430.654 14.4 430.882 14.016 431.266 13.032L434.086 5.88H433.042L431.218 10.764L429.334 5.88H428.29L430.678 12.036ZM438.576 12.072C440.472 12.072 441.864 10.992 442.176 9.252H441.072C440.76 10.428 439.86 11.112 438.564 11.112C436.8 11.112 435.72 9.732 435.72 7.8C435.72 5.856 436.848 4.488 438.612 4.488C439.872 4.488 440.76 5.184 441.084 6.348H442.188C441.876 4.608 440.52 3.528 438.66 3.528C436.224 3.528 434.652 5.256 434.652 7.8C434.652 10.368 436.164 12.072 438.576 12.072ZM443.437 12H444.481V8.772H446.593L448.369 12H449.545L447.661 8.568C448.741 8.268 449.449 7.356 449.449 6.18C449.449 4.608 448.333 3.6 446.653 3.6H443.437V12ZM446.581 4.56C447.673 4.56 448.381 5.196 448.381 6.18C448.381 7.164 447.673 7.812 446.569 7.812H444.481V4.56H446.581ZM452.948 12H454.208L457.232 3.6H456.152L453.596 10.764L451.028 3.6H449.924L452.948 12ZM463.612 12H464.872L467.896 3.6H466.816L464.26 10.764L461.692 3.6H460.588L463.612 12ZM470.851 12.072C472.639 12.072 473.887 10.8 473.887 8.94C473.887 7.08 472.639 5.808 470.851 5.808C469.063 5.808 467.815 7.08 467.815 8.94C467.815 10.8 469.063 12.072 470.851 12.072ZM470.851 11.196C469.639 11.196 468.823 10.248 468.823 8.94C468.823 7.632 469.639 6.684 470.851 6.684C472.063 6.684 472.879 7.632 472.879 8.94C472.879 10.248 472.063 11.196 470.851 11.196ZM475.401 10.404C475.401 11.484 475.821 12 476.997 12H478.065V11.1H477.093C476.565 11.1 476.385 10.908 476.385 10.392V6.78H477.969V5.88H476.385V4.164H475.401V5.88H474.297V6.78H475.401V10.404ZM481.586 12.072C483.002 12.072 484.01 11.172 484.25 9.948H483.254C483.05 10.74 482.378 11.196 481.514 11.196C480.35 11.196 479.618 10.38 479.582 9.216V9.132H484.334C484.346 8.988 484.358 8.844 484.358 8.712C484.31 6.9 483.182 5.808 481.454 5.808C479.726 5.808 478.55 7.08 478.55 8.94C478.55 10.812 479.75 12.072 481.586 12.072ZM479.642 8.316C479.738 7.32 480.566 6.672 481.466 6.672C482.486 6.672 483.206 7.26 483.302 8.316H479.642ZM491.453 12C494.081 12 495.677 10.308 495.677 7.8C495.677 5.28 494.081 3.6 491.453 3.6H488.765V12H491.453ZM489.809 4.56H491.453C493.493 4.56 494.609 5.832 494.609 7.8C494.609 9.756 493.493 11.04 491.453 11.04H489.809V4.56ZM499.668 12.072C501.084 12.072 502.092 11.172 502.332 9.948H501.336C501.132 10.74 500.46 11.196 499.596 11.196C498.432 11.196 497.7 10.38 497.664 9.216V9.132H502.416C502.428 8.988 502.44 8.844 502.44 8.712C502.392 6.9 501.264 5.808 499.536 5.808C497.808 5.808 496.632 7.08 496.632 8.94C496.632 10.812 497.832 12.072 499.668 12.072ZM497.724 8.316C497.82 7.32 498.648 6.672 499.548 6.672C500.568 6.672 501.288 7.26 501.384 8.316H497.724ZM508.975 11.1C508.699 11.1 508.579 11.016 508.579 10.692V8.016C508.579 6.588 507.679 5.808 506.119 5.808C504.643 5.808 503.671 6.516 503.527 7.716H504.511C504.631 7.092 505.207 6.684 506.071 6.684C507.031 6.684 507.595 7.164 507.595 7.956V8.376H505.795C504.187 8.376 503.335 9.06 503.335 10.284C503.335 11.4 504.247 12.072 505.591 12.072C506.647 12.072 507.259 11.604 507.643 11.052C507.655 11.652 507.931 12 508.747 12H509.287V11.1H508.975ZM507.595 9.504C507.595 10.524 506.887 11.232 505.663 11.232C504.847 11.232 504.331 10.824 504.331 10.224C504.331 9.528 504.823 9.204 505.699 9.204H507.595V9.504ZM514.797 6.816C514.425 6.228 513.729 5.808 512.697 5.808C511.005 5.808 509.757 7.068 509.757 8.94C509.757 10.932 511.005 12.072 512.697 12.072C513.729 12.072 514.389 11.58 514.797 11.016L514.929 12H515.781V3.6H514.797V6.816ZM512.793 11.196C511.581 11.196 510.765 10.272 510.765 8.94C510.765 7.608 511.581 6.684 512.793 6.684C513.993 6.684 514.809 7.62 514.809 8.964C514.809 10.272 513.993 11.196 512.793 11.196ZM517.121 12H518.105V3.6H517.121V12ZM519.957 4.848C520.317 4.848 520.593 4.584 520.593 4.212C520.593 3.84 520.317 3.576 519.957 3.576C519.597 3.576 519.321 3.84 519.321 4.212C519.321 4.584 519.597 4.848 519.957 4.848ZM519.453 12H520.437V5.88H519.453V12ZM524.785 5.808C523.753 5.808 523.129 6.24 522.769 6.708L522.637 5.88H521.785V12H522.769V8.94C522.769 7.56 523.465 6.708 524.617 6.708C525.685 6.708 526.273 7.38 526.273 8.652V12H527.257V8.592C527.257 6.564 526.165 5.808 524.785 5.808ZM531.262 12.072C532.678 12.072 533.686 11.172 533.926 9.948H532.93C532.726 10.74 532.054 11.196 531.19 11.196C530.026 11.196 529.294 10.38 529.258 9.216V9.132H534.01C534.022 8.988 534.034 8.844 534.034 8.712C533.986 6.9 532.858 5.808 531.13 5.808C529.402 5.808 528.226 7.08 528.226 8.94C528.226 10.812 529.426 12.072 531.262 12.072ZM529.318 8.316C529.414 7.32 530.242 6.672 531.142 6.672C532.162 6.672 532.882 7.26 532.978 8.316H529.318Z'
-				}
-				fill={'currentcolor'}
-			/>
-			<path
-				d={
-					'M852.64 12H853.888L855.796 5.34L857.716 12H858.964L861.424 3.6H860.344L858.364 10.656L856.372 3.6H855.316L853.252 10.656L851.284 3.6H850.192L852.64 12ZM865.285 5.808C864.253 5.808 863.629 6.312 863.269 6.888V3.6H862.285V12H863.269V8.94C863.269 7.56 863.977 6.708 865.117 6.708C866.173 6.708 866.785 7.38 866.785 8.652V12H867.769V8.592C867.769 6.564 866.653 5.808 865.285 5.808ZM871.773 12.072C873.189 12.072 874.197 11.172 874.437 9.948H873.441C873.237 10.74 872.565 11.196 871.701 11.196C870.537 11.196 869.805 10.38 869.769 9.216V9.132H874.521C874.533 8.988 874.545 8.844 874.545 8.712C874.497 6.9 873.369 5.808 871.641 5.808C869.913 5.808 868.737 7.08 868.737 8.94C868.737 10.812 869.937 12.072 871.773 12.072ZM869.829 8.316C869.925 7.32 870.753 6.672 871.653 6.672C872.673 6.672 873.393 7.26 873.489 8.316H869.829ZM878.668 5.808C877.636 5.808 877.012 6.24 876.652 6.708L876.52 5.88H875.668V12H876.652V8.94C876.652 7.56 877.348 6.708 878.5 6.708C879.568 6.708 880.156 7.38 880.156 8.652V12H881.14V8.592C881.14 6.564 880.048 5.808 878.668 5.808ZM887.311 12.036L887.011 12.78C886.783 13.356 886.687 13.5 886.123 13.5H885.247V14.4H886.423C887.287 14.4 887.515 14.016 887.899 13.032L890.719 5.88H889.675L887.851 10.764L885.967 5.88H884.923L887.311 12.036ZM894.004 12.072C895.792 12.072 897.04 10.8 897.04 8.94C897.04 7.08 895.792 5.808 894.004 5.808C892.216 5.808 890.968 7.08 890.968 8.94C890.968 10.8 892.216 12.072 894.004 12.072ZM894.004 11.196C892.792 11.196 891.976 10.248 891.976 8.94C891.976 7.632 892.792 6.684 894.004 6.684C895.216 6.684 896.032 7.632 896.032 8.94C896.032 10.248 895.216 11.196 894.004 11.196ZM902.452 8.94C902.452 10.32 901.78 11.172 900.664 11.172C899.644 11.172 899.068 10.5 899.068 9.228V5.88H898.084V9.288C898.084 11.316 899.152 12.072 900.496 12.072C901.504 12.072 902.092 11.64 902.452 11.184L902.584 12H903.436V5.88H902.452V8.94ZM910.666 12.072C912.154 12.072 913.15 11.232 913.39 9.912H912.346C912.154 10.692 911.554 11.196 910.678 11.196C909.502 11.196 908.71 10.284 908.71 8.952C908.71 7.62 909.502 6.684 910.678 6.684C911.578 6.684 912.178 7.176 912.358 7.992H913.39C913.138 6.648 912.19 5.808 910.69 5.808C908.914 5.808 907.702 7.08 907.702 8.952C907.702 10.86 908.878 12.072 910.666 12.072ZM919.881 11.1C919.605 11.1 919.485 11.016 919.485 10.692V8.016C919.485 6.588 918.585 5.808 917.025 5.808C915.549 5.808 914.577 6.516 914.433 7.716H915.417C915.537 7.092 916.113 6.684 916.977 6.684C917.937 6.684 918.501 7.164 918.501 7.956V8.376H916.701C915.093 8.376 914.241 9.06 914.241 10.284C914.241 11.4 915.153 12.072 916.497 12.072C917.553 12.072 918.165 11.604 918.549 11.052C918.561 11.652 918.837 12 919.653 12H920.193V11.1H919.881ZM918.501 9.504C918.501 10.524 917.793 11.232 916.569 11.232C915.753 11.232 915.237 10.824 915.237 10.224C915.237 9.528 915.729 9.204 916.605 9.204H918.501V9.504ZM923.949 5.808C922.917 5.808 922.293 6.24 921.933 6.708L921.801 5.88H920.949V12H921.933V8.94C921.933 7.56 922.629 6.708 923.781 6.708C924.849 6.708 925.437 7.38 925.437 8.652V12H926.421V8.592C926.421 6.564 925.329 5.808 923.949 5.808ZM932.221 12H933.301L934.837 7.452L936.313 12H937.393L939.433 5.88H938.437L936.841 10.86L935.317 5.88H934.405L932.797 10.86L931.273 5.88H930.241L932.221 12ZM940.707 4.848C941.067 4.848 941.343 4.584 941.343 4.212C941.343 3.84 941.067 3.576 940.707 3.576C940.347 3.576 940.071 3.84 940.071 4.212C940.071 4.584 940.347 4.848 940.707 4.848ZM940.203 12H941.187V5.88H940.203V12ZM942.991 10.404C942.991 11.484 943.411 12 944.587 12H945.655V11.1H944.683C944.155 11.1 943.975 10.908 943.975 10.392V6.78H945.559V5.88H943.975V4.164H942.991V5.88H941.887V6.78H942.991V10.404ZM949.508 5.808C948.476 5.808 947.852 6.312 947.492 6.888V3.6H946.508V12H947.492V8.94C947.492 7.56 948.2 6.708 949.34 6.708C950.396 6.708 951.008 7.38 951.008 8.652V12H951.992V8.592C951.992 6.564 950.876 5.808 949.508 5.808ZM958 6.816C957.628 6.228 956.932 5.808 955.9 5.808C954.208 5.808 952.96 7.068 952.96 8.94C952.96 10.932 954.208 12.072 955.9 12.072C956.932 12.072 957.592 11.58 958 11.016L958.132 12H958.984V3.6H958V6.816ZM955.996 11.196C954.784 11.196 953.968 10.272 953.968 8.94C953.968 7.608 954.784 6.684 955.996 6.684C957.196 6.684 958.012 7.62 958.012 8.964C958.012 10.272 957.196 11.196 955.996 11.196ZM963.156 5.88C962.064 5.88 961.596 6.336 961.308 6.804L961.176 5.88H960.324V12H961.308V8.88C961.308 7.908 961.656 6.84 962.94 6.84H963.432V5.88H963.156ZM969.486 11.1C969.21 11.1 969.09 11.016 969.09 10.692V8.016C969.09 6.588 968.19 5.808 966.63 5.808C965.154 5.808 964.182 6.516 964.038 7.716H965.022C965.142 7.092 965.718 6.684 966.582 6.684C967.542 6.684 968.106 7.164 968.106 7.956V8.376H966.306C964.698 8.376 963.846 9.06 963.846 10.284C963.846 11.4 964.758 12.072 966.102 12.072C967.158 12.072 967.77 11.604 968.154 11.052C968.166 11.652 968.442 12 969.258 12H969.798V11.1H969.486ZM968.106 9.504C968.106 10.524 967.398 11.232 966.174 11.232C965.358 11.232 964.842 10.824 964.842 10.224C964.842 9.528 965.334 9.204 966.21 9.204H968.106V9.504ZM971.795 12H972.875L974.411 7.452L975.887 12H976.967L979.007 5.88H978.011L976.415 10.86L974.891 5.88H973.979L972.371 10.86L970.847 5.88H969.815L971.795 12Z'
-				}
-				fill={'currentcolor'}
-			/>
-			<path
-				d={
-					'M280.428 109.072C282.324 109.072 283.716 107.992 284.028 106.252H282.924C282.612 107.428 281.712 108.112 280.416 108.112C278.652 108.112 277.572 106.732 277.572 104.8C277.572 102.856 278.7 101.488 280.464 101.488C281.724 101.488 282.612 102.184 282.936 103.348H284.04C283.728 101.608 282.372 100.528 280.512 100.528C278.076 100.528 276.504 102.256 276.504 104.8C276.504 107.368 278.016 109.072 280.428 109.072ZM289.429 105.94C289.429 107.32 288.757 108.172 287.641 108.172C286.621 108.172 286.045 107.5 286.045 106.228V102.88H285.061V106.288C285.061 108.316 286.129 109.072 287.473 109.072C288.481 109.072 289.069 108.64 289.429 108.184L289.561 109H290.413V102.88H289.429V105.94ZM294.586 102.88C293.494 102.88 293.026 103.336 292.738 103.804L292.606 102.88H291.754V109H292.738V105.88C292.738 104.908 293.086 103.84 294.37 103.84H294.862V102.88H294.586ZM298.547 102.88C297.455 102.88 296.987 103.336 296.699 103.804L296.567 102.88H295.715V109H296.699V105.88C296.699 104.908 297.047 103.84 298.331 103.84H298.823V102.88H298.547ZM302.144 109.072C303.56 109.072 304.568 108.172 304.808 106.948H303.812C303.608 107.74 302.936 108.196 302.072 108.196C300.908 108.196 300.176 107.38 300.14 106.216V106.132H304.892C304.904 105.988 304.916 105.844 304.916 105.712C304.868 103.9 303.74 102.808 302.012 102.808C300.284 102.808 299.108 104.08 299.108 105.94C299.108 107.812 300.308 109.072 302.144 109.072ZM300.2 105.316C300.296 104.32 301.124 103.672 302.024 103.672C303.044 103.672 303.764 104.26 303.86 105.316H300.2ZM309.039 102.808C308.007 102.808 307.383 103.24 307.023 103.708L306.891 102.88H306.039V109H307.023V105.94C307.023 104.56 307.719 103.708 308.871 103.708C309.939 103.708 310.527 104.38 310.527 105.652V109H311.511V105.592C311.511 103.564 310.419 102.808 309.039 102.808ZM313.222 107.404C313.222 108.484 313.642 109 314.818 109H315.886V108.1H314.914C314.386 108.1 314.206 107.908 314.206 107.392V103.78H315.79V102.88H314.206V101.164H313.222V102.88H312.118V103.78H313.222V107.404ZM320.023 109H321.067V105.772H323.251C324.931 105.772 326.035 104.752 326.035 103.18C326.035 101.608 324.931 100.6 323.251 100.6H320.023V109ZM323.167 101.56C324.271 101.56 324.967 102.184 324.967 103.18C324.967 104.176 324.259 104.812 323.155 104.812H321.067V101.56H323.167ZM329.508 109.072C330.924 109.072 331.932 108.172 332.172 106.948H331.176C330.972 107.74 330.3 108.196 329.436 108.196C328.272 108.196 327.54 107.38 327.504 106.216V106.132H332.256C332.268 105.988 332.28 105.844 332.28 105.712C332.232 103.9 331.104 102.808 329.376 102.808C327.648 102.808 326.472 104.08 326.472 105.94C326.472 107.812 327.672 109.072 329.508 109.072ZM327.564 105.316C327.66 104.32 328.488 103.672 329.388 103.672C330.408 103.672 331.128 104.26 331.224 105.316H327.564ZM336.234 102.88C335.142 102.88 334.674 103.336 334.386 103.804L334.254 102.88H333.402V109H334.386V105.88C334.386 104.908 334.734 103.84 336.018 103.84H336.51V102.88H336.234ZM337.867 101.848C338.227 101.848 338.503 101.584 338.503 101.212C338.503 100.84 338.227 100.576 337.867 100.576C337.507 100.576 337.231 100.84 337.231 101.212C337.231 101.584 337.507 101.848 337.867 101.848ZM337.363 109H338.347V102.88H337.363V109ZM342.515 109.072C344.303 109.072 345.551 107.8 345.551 105.94C345.551 104.08 344.303 102.808 342.515 102.808C340.727 102.808 339.479 104.08 339.479 105.94C339.479 107.8 340.727 109.072 342.515 109.072ZM342.515 108.196C341.303 108.196 340.487 107.248 340.487 105.94C340.487 104.632 341.303 103.684 342.515 103.684C343.727 103.684 344.543 104.632 344.543 105.94C344.543 107.248 343.727 108.196 342.515 108.196ZM351.504 103.816C351.132 103.228 350.436 102.808 349.404 102.808C347.712 102.808 346.464 104.068 346.464 105.94C346.464 107.932 347.712 109.072 349.404 109.072C350.436 109.072 351.096 108.58 351.504 108.016L351.636 109H352.488V100.6H351.504V103.816ZM349.5 108.196C348.288 108.196 347.472 107.272 347.472 105.94C347.472 104.608 348.288 103.684 349.5 103.684C350.7 103.684 351.516 104.62 351.516 105.964C351.516 107.272 350.7 108.196 349.5 108.196Z'
-				}
-				fill={'black'}
-			/>
-			<path
-				d={
-					'M490.964 75.6765L490.22 75.3765C489.644 75.1485 489.5 75.0525 489.5 74.4885L489.5 73.6125L488.6 73.6125L488.6 74.7885C488.6 75.6525 488.984 75.8805 489.968 76.2645L497.12 79.0845L497.12 78.0405L492.236 76.2165L497.12 74.3325L497.12 73.2885L490.964 75.6765ZM490.928 83.5745C490.928 85.4705 492.008 86.8625 493.748 87.1745L493.748 86.0705C492.572 85.7585 491.888 84.8585 491.888 83.5625C491.888 81.7985 493.268 80.7185 495.2 80.7185C497.144 80.7185 498.512 81.8465 498.512 83.6105C498.512 84.8705 497.816 85.7585 496.652 86.0825L496.652 87.1865C498.392 86.8745 499.472 85.5185 499.472 83.6585C499.472 81.2225 497.744 79.6505 495.2 79.6505C492.632 79.6505 490.928 81.1625 490.928 83.5745ZM491 88.4351L491 89.4791L494.228 89.4791L494.228 91.5911L491 93.3671L491 94.5431L494.432 92.6591C494.732 93.7391 495.644 94.4471 496.82 94.4471C498.392 94.4471 499.4 93.3311 499.4 91.6511L499.4 88.4351L491 88.4351ZM498.44 91.5791C498.44 92.6711 497.804 93.3791 496.82 93.3791C495.836 93.3791 495.188 92.6711 495.188 91.5671L495.188 89.4791L498.44 89.4791L498.44 91.5791ZM491 97.9465L491 99.2065L499.4 102.23L499.4 101.15L492.236 98.5945L499.4 96.0265L499.4 94.9225L491 97.9465ZM491 108.611L491 109.871L499.4 112.895L499.4 111.815L492.236 109.259L499.4 106.691L499.4 105.587L491 108.611ZM490.928 115.849C490.928 117.637 492.2 118.885 494.06 118.885C495.92 118.885 497.192 117.637 497.192 115.849C497.192 114.061 495.92 112.813 494.06 112.813C492.2 112.813 490.928 114.061 490.928 115.849ZM491.804 115.849C491.804 114.637 492.752 113.821 494.06 113.821C495.368 113.821 496.316 114.637 496.316 115.849C496.316 117.061 495.368 117.877 494.06 117.877C492.752 117.877 491.804 117.061 491.804 115.849ZM492.596 120.399C491.516 120.399 491 120.819 491 121.995L491 123.063L491.9 123.063L491.9 122.091C491.9 121.563 492.092 121.383 492.608 121.383L496.22 121.383L496.22 122.967L497.12 122.967L497.12 121.383L498.836 121.383L498.836 120.399L497.12 120.399L497.12 119.295L496.22 119.295L496.22 120.399L492.596 120.399ZM490.928 126.584C490.928 128 491.828 129.008 493.052 129.248L493.052 128.252C492.26 128.048 491.804 127.376 491.804 126.512C491.804 125.348 492.62 124.616 493.784 124.58L493.868 124.58L493.868 129.332C494.012 129.344 494.156 129.356 494.288 129.356C496.1 129.308 497.192 128.18 497.192 126.452C497.192 124.724 495.92 123.548 494.06 123.548C492.188 123.548 490.928 124.748 490.928 126.584ZM494.684 124.64C495.68 124.736 496.328 125.564 496.328 126.464C496.328 127.484 495.74 128.204 494.684 128.3L494.684 124.64ZM493.016 130.203C491.672 130.227 490.928 131.319 490.928 132.903C490.928 134.259 491.636 135.351 492.776 135.351C494.132 135.351 494.396 134.235 494.564 132.843C494.672 131.955 494.756 131.307 495.428 131.307C495.992 131.307 496.352 131.883 496.352 132.735C496.352 133.575 495.884 134.151 495.26 134.223L495.26 135.231C496.532 135.135 497.18 134.127 497.192 132.735C497.204 131.355 496.472 130.335 495.356 130.335C494.084 130.335 493.808 131.427 493.64 132.795C493.532 133.707 493.472 134.379 492.704 134.379C492.14 134.379 491.768 133.767 491.768 132.903C491.768 131.883 492.296 131.247 493.016 131.211L493.016 130.203Z'
-				}
-				fill={'black'}
-			/>
-			<path
-				d={
-					'M198.64 188H199.888L201.796 181.34L203.716 188H204.964L207.424 179.6H206.344L204.364 186.656L202.372 179.6H201.316L199.252 186.656L197.284 179.6H196.192L198.64 188ZM210.578 188.072C211.994 188.072 213.002 187.172 213.242 185.948H212.246C212.042 186.74 211.37 187.196 210.506 187.196C209.342 187.196 208.61 186.38 208.574 185.216V185.132H213.326C213.338 184.988 213.35 184.844 213.35 184.712C213.302 182.9 212.174 181.808 210.446 181.808C208.718 181.808 207.542 183.08 207.542 184.94C207.542 186.812 208.742 188.072 210.578 188.072ZM208.634 184.316C208.73 183.32 209.558 182.672 210.458 182.672C211.478 182.672 212.198 183.26 212.294 184.316H208.634ZM217.293 188.072C218.709 188.072 219.717 187.172 219.957 185.948H218.961C218.757 186.74 218.085 187.196 217.221 187.196C216.057 187.196 215.325 186.38 215.289 185.216V185.132H220.041C220.053 184.988 220.065 184.844 220.065 184.712C220.017 182.9 218.889 181.808 217.161 181.808C215.433 181.808 214.257 183.08 214.257 184.94C214.257 186.812 215.457 188.072 217.293 188.072ZM215.349 184.316C215.445 183.32 216.273 182.672 217.173 182.672C218.193 182.672 218.913 183.26 219.009 184.316H215.349ZM221.188 188H222.172V186.392L223.408 185.12L225.436 188H226.624L224.068 184.412L226.456 181.88H225.244L222.172 185.132V179.6H221.188V188ZM231.789 180.512C231.789 180.932 231.597 181.232 231.201 181.232H229.833V182.132H231.189C231.453 182.132 231.645 182.084 231.765 181.976V188H232.773V179.6H231.789V180.512Z'
-				}
-				fill={'black'}
-			/>
-			<path
-				d={
-					'M397.64 188H398.888L400.796 181.34L402.716 188H403.964L406.424 179.6H405.344L403.364 186.656L401.372 179.6H400.316L398.252 186.656L396.284 179.6H395.192L397.64 188ZM409.578 188.072C410.994 188.072 412.002 187.172 412.242 185.948H411.246C411.042 186.74 410.37 187.196 409.506 187.196C408.342 187.196 407.61 186.38 407.574 185.216V185.132H412.326C412.338 184.988 412.35 184.844 412.35 184.712C412.302 182.9 411.174 181.808 409.446 181.808C407.718 181.808 406.542 183.08 406.542 184.94C406.542 186.812 407.742 188.072 409.578 188.072ZM407.634 184.316C407.73 183.32 408.558 182.672 409.458 182.672C410.478 182.672 411.198 183.26 411.294 184.316H407.634ZM416.293 188.072C417.709 188.072 418.717 187.172 418.957 185.948H417.961C417.757 186.74 417.085 187.196 416.221 187.196C415.057 187.196 414.325 186.38 414.289 185.216V185.132H419.041C419.053 184.988 419.065 184.844 419.065 184.712C419.017 182.9 417.889 181.808 416.161 181.808C414.433 181.808 413.257 183.08 413.257 184.94C413.257 186.812 414.457 188.072 416.293 188.072ZM414.349 184.316C414.445 183.32 415.273 182.672 416.173 182.672C417.193 182.672 417.913 183.26 418.009 184.316H414.349ZM420.188 188H421.172V186.392L422.408 185.12L424.436 188H425.624L423.068 184.412L425.456 181.88H424.244L421.172 185.132V179.6H420.188V188ZM429.049 188H434.833V187.076H430.441L432.961 184.94C433.957 184.088 434.773 183.296 434.773 182.036C434.773 180.524 433.717 179.528 431.905 179.528C430.213 179.528 429.097 180.656 429.097 182.24V182.456H430.105V182.3C430.105 181.148 430.825 180.428 431.905 180.428C433.021 180.428 433.741 181.028 433.741 182.084C433.741 182.996 433.093 183.644 432.313 184.316L429.049 187.148V188Z'
-				}
-				fill={'black'}
-			/>
-			<path
-				d={
-					'M597.64 188H598.888L600.796 181.34L602.716 188H603.964L606.424 179.6H605.344L603.364 186.656L601.372 179.6H600.316L598.252 186.656L596.284 179.6H595.192L597.64 188ZM609.578 188.072C610.994 188.072 612.002 187.172 612.242 185.948H611.246C611.042 186.74 610.37 187.196 609.506 187.196C608.342 187.196 607.61 186.38 607.574 185.216V185.132H612.326C612.338 184.988 612.35 184.844 612.35 184.712C612.302 182.9 611.174 181.808 609.446 181.808C607.718 181.808 606.542 183.08 606.542 184.94C606.542 186.812 607.742 188.072 609.578 188.072ZM607.634 184.316C607.73 183.32 608.558 182.672 609.458 182.672C610.478 182.672 611.198 183.26 611.294 184.316H607.634ZM616.293 188.072C617.709 188.072 618.717 187.172 618.957 185.948H617.961C617.757 186.74 617.085 187.196 616.221 187.196C615.057 187.196 614.325 186.38 614.289 185.216V185.132H619.041C619.053 184.988 619.065 184.844 619.065 184.712C619.017 182.9 617.889 181.808 616.161 181.808C614.433 181.808 613.257 183.08 613.257 184.94C613.257 186.812 614.457 188.072 616.293 188.072ZM614.349 184.316C614.445 183.32 615.273 182.672 616.173 182.672C617.193 182.672 617.913 183.26 618.009 184.316H614.349ZM620.188 188H621.172V186.392L622.408 185.12L624.436 188H625.624L623.068 184.412L625.456 181.88H624.244L621.172 185.132V179.6H620.188V188ZM633.573 183.632C634.377 183.416 634.929 182.804 634.929 181.748C634.929 180.428 633.837 179.528 632.109 179.528C630.453 179.528 629.301 180.548 629.289 182.024H630.285C630.297 181.064 631.065 180.428 632.121 180.428C633.129 180.428 633.897 180.956 633.897 181.808C633.897 182.66 633.273 183.284 632.025 183.284H631.365V184.16H632.085C633.393 184.16 634.113 184.724 634.113 185.6C634.113 186.56 633.273 187.172 632.181 187.172C630.993 187.172 630.129 186.512 630.117 185.408H629.121C629.121 187.076 630.297 188.072 632.121 188.072C633.945 188.072 635.145 187.1 635.145 185.672C635.145 184.388 634.341 183.776 633.573 183.632Z'
-				}
-				fill={'black'}
-			/>
-			<path
-				d={
-					'M797.64 188H798.888L800.796 181.34L802.716 188H803.964L806.424 179.6H805.344L803.364 186.656L801.372 179.6H800.316L798.252 186.656L796.284 179.6H795.192L797.64 188ZM809.578 188.072C810.994 188.072 812.002 187.172 812.242 185.948H811.246C811.042 186.74 810.37 187.196 809.506 187.196C808.342 187.196 807.61 186.38 807.574 185.216V185.132H812.326C812.338 184.988 812.35 184.844 812.35 184.712C812.302 182.9 811.174 181.808 809.446 181.808C807.718 181.808 806.542 183.08 806.542 184.94C806.542 186.812 807.742 188.072 809.578 188.072ZM807.634 184.316C807.73 183.32 808.558 182.672 809.458 182.672C810.478 182.672 811.198 183.26 811.294 184.316H807.634ZM816.293 188.072C817.709 188.072 818.717 187.172 818.957 185.948H817.961C817.757 186.74 817.085 187.196 816.221 187.196C815.057 187.196 814.325 186.38 814.289 185.216V185.132H819.041C819.053 184.988 819.065 184.844 819.065 184.712C819.017 182.9 817.889 181.808 816.161 181.808C814.433 181.808 813.257 183.08 813.257 184.94C813.257 186.812 814.457 188.072 816.293 188.072ZM814.349 184.316C814.445 183.32 815.273 182.672 816.173 182.672C817.193 182.672 817.913 183.26 818.009 184.316H814.349ZM820.188 188H821.172V186.392L822.408 185.12L824.436 188H825.624L823.068 184.412L825.456 181.88H824.244L821.172 185.132V179.6H820.188V188ZM828.941 185.948H833.345V188H834.353V185.948H835.709V185.024H834.353V179.6H833.225L828.941 185.048V185.948ZM833.345 180.836V185.024H830.069L833.345 180.836Z'
-				}
-				fill={'black'}
-			/>
-			<path
-				d={
-					'M685.836 109V102.52L690.348 109H691.308V100.6H690.288V107.08L685.776 100.6H684.816V109H685.836ZM695.613 109.072C697.029 109.072 698.037 108.172 698.277 106.948H697.281C697.077 107.74 696.405 108.196 695.541 108.196C694.377 108.196 693.645 107.38 693.609 106.216V106.132H698.361C698.373 105.988 698.385 105.844 698.385 105.712C698.337 103.9 697.209 102.808 695.481 102.808C693.753 102.808 692.577 104.08 692.577 105.94C692.577 107.812 693.777 109.072 695.613 109.072ZM693.669 105.316C693.765 104.32 694.593 103.672 695.493 103.672C696.513 103.672 697.233 104.26 697.329 105.316H693.669ZM698.602 109H699.718L701.506 106.54L703.294 109H704.47L702.13 105.856L704.314 102.88H703.198L701.554 105.148L699.934 102.88H698.758L700.93 105.856L698.602 109ZM705.683 107.404C705.683 108.484 706.103 109 707.279 109H708.347V108.1H707.375C706.847 108.1 706.667 107.908 706.667 107.392V103.78H708.251V102.88H706.667V101.164H705.683V102.88H704.579V103.78H705.683V107.404ZM712.484 109H713.528V105.772H715.712C717.392 105.772 718.496 104.752 718.496 103.18C718.496 101.608 717.392 100.6 715.712 100.6H712.484V109ZM715.628 101.56C716.732 101.56 717.428 102.184 717.428 103.18C717.428 104.176 716.72 104.812 715.616 104.812H713.528V101.56H715.628ZM721.969 109.072C723.385 109.072 724.393 108.172 724.633 106.948H723.637C723.433 107.74 722.761 108.196 721.897 108.196C720.733 108.196 720.001 107.38 719.965 106.216V106.132H724.717C724.729 105.988 724.741 105.844 724.741 105.712C724.693 103.9 723.565 102.808 721.837 102.808C720.109 102.808 718.933 104.08 718.933 105.94C718.933 107.812 720.133 109.072 721.969 109.072ZM720.025 105.316C720.121 104.32 720.949 103.672 721.849 103.672C722.869 103.672 723.589 104.26 723.685 105.316H720.025ZM728.695 102.88C727.603 102.88 727.135 103.336 726.847 103.804L726.715 102.88H725.863V109H726.847V105.88C726.847 104.908 727.195 103.84 728.479 103.84H728.971V102.88H728.695ZM730.328 101.848C730.688 101.848 730.964 101.584 730.964 101.212C730.964 100.84 730.688 100.576 730.328 100.576C729.968 100.576 729.692 100.84 729.692 101.212C729.692 101.584 729.968 101.848 730.328 101.848ZM729.824 109H730.808V102.88H729.824V109ZM734.976 109.072C736.764 109.072 738.012 107.8 738.012 105.94C738.012 104.08 736.764 102.808 734.976 102.808C733.188 102.808 731.94 104.08 731.94 105.94C731.94 107.8 733.188 109.072 734.976 109.072ZM734.976 108.196C733.764 108.196 732.948 107.248 732.948 105.94C732.948 104.632 733.764 103.684 734.976 103.684C736.188 103.684 737.004 104.632 737.004 105.94C737.004 107.248 736.188 108.196 734.976 108.196ZM743.965 103.816C743.593 103.228 742.897 102.808 741.865 102.808C740.173 102.808 738.925 104.068 738.925 105.94C738.925 107.932 740.173 109.072 741.865 109.072C742.897 109.072 743.557 108.58 743.965 108.016L744.097 109H744.949V100.6H743.965V103.816ZM741.961 108.196C740.749 108.196 739.933 107.272 739.933 105.94C739.933 104.608 740.749 103.684 741.961 103.684C743.161 103.684 743.977 104.62 743.977 105.964C743.977 107.272 743.161 108.196 741.961 108.196Z'
-				}
-				fill={'black'}
-			/>
-			<path
-				d={
-					'M2.352 100H3.54L5.832 93.88H4.8L2.964 98.884L1.116 93.88H0.06L2.352 100ZM6.56653 100H7.55053V91.6H6.56653V100ZM12.8466 97V96.172H8.85056V97H12.8466ZM15.6785 100.036L15.3785 100.78C15.1505 101.356 15.0545 101.5 14.4905 101.5H13.6145V102.4H14.7905C15.6545 102.4 15.8825 102.016 16.2665 101.032L19.0865 93.88H18.0425L16.2185 98.764L14.3345 93.88H13.2905L15.6785 100.036ZM23.5764 100.072C25.4724 100.072 26.8644 98.992 27.1764 97.252H26.0724C25.7604 98.428 24.8604 99.112 23.5644 99.112C21.8004 99.112 20.7204 97.732 20.7204 95.8C20.7204 93.856 21.8484 92.488 23.6124 92.488C24.8724 92.488 25.7604 93.184 26.0844 94.348H27.1884C26.8764 92.608 25.5204 91.528 23.6604 91.528C21.2244 91.528 19.6524 93.256 19.6524 95.8C19.6524 98.368 21.1644 100.072 23.5764 100.072ZM28.4371 100H29.4811V96.772H31.5931L33.3691 100H34.5451L32.6611 96.568C33.7411 96.268 34.4491 95.356 34.4491 94.18C34.4491 92.608 33.3331 91.6 31.6531 91.6H28.4371V100ZM31.5811 92.56C32.6731 92.56 33.3811 93.196 33.3811 94.18C33.3811 95.164 32.6731 95.812 31.5691 95.812H29.4811V92.56H31.5811ZM37.9484 100H39.2084L42.2324 91.6H41.1524L38.5964 98.764L36.0284 91.6H34.9244L37.9484 100ZM46.2965 100H47.3405V96.772H49.5245C51.2045 96.772 52.3085 95.752 52.3085 94.18C52.3085 92.608 51.2045 91.6 49.5245 91.6H46.2965V100ZM49.4405 92.56C50.5445 92.56 51.2405 93.184 51.2405 94.18C51.2405 95.176 50.5325 95.812 49.4285 95.812H47.3405V92.56H49.4405ZM55.7811 100.072C57.1971 100.072 58.2051 99.172 58.4451 97.948H57.4491C57.2451 98.74 56.5731 99.196 55.7091 99.196C54.5451 99.196 53.8131 98.38 53.7771 97.216V97.132H58.5291C58.5411 96.988 58.5531 96.844 58.5531 96.712C58.5051 94.9 57.3771 93.808 55.6491 93.808C53.9211 93.808 52.7451 95.08 52.7451 96.94C52.7451 98.812 53.9451 100.072 55.7811 100.072ZM53.8371 96.316C53.9331 95.32 54.7611 94.672 55.6611 94.672C56.6811 94.672 57.4011 95.26 57.4971 96.316H53.8371ZM62.5079 93.88C61.4159 93.88 60.9479 94.336 60.6599 94.804L60.5279 93.88H59.6759V100H60.6599V96.88C60.6599 95.908 61.0079 94.84 62.2919 94.84H62.7839V93.88H62.5079ZM64.1408 92.848C64.5008 92.848 64.7768 92.584 64.7768 92.212C64.7768 91.84 64.5008 91.576 64.1408 91.576C63.7808 91.576 63.5048 91.84 63.5048 92.212C63.5048 92.584 63.7808 92.848 64.1408 92.848ZM63.6368 100H64.6208V93.88H63.6368V100ZM68.7889 100.072C70.5769 100.072 71.8249 98.8 71.8249 96.94C71.8249 95.08 70.5769 93.808 68.7889 93.808C67.0009 93.808 65.7529 95.08 65.7529 96.94C65.7529 98.8 67.0009 100.072 68.7889 100.072ZM68.7889 99.196C67.5769 99.196 66.7609 98.248 66.7609 96.94C66.7609 95.632 67.5769 94.684 68.7889 94.684C70.0009 94.684 70.8169 95.632 70.8169 96.94C70.8169 98.248 70.0009 99.196 68.7889 99.196ZM77.7773 94.816C77.4053 94.228 76.7093 93.808 75.6773 93.808C73.9853 93.808 72.7373 95.068 72.7373 96.94C72.7373 98.932 73.9853 100.072 75.6773 100.072C76.7093 100.072 77.3693 99.58 77.7773 99.016L77.9093 100H78.7613V91.6H77.7773V94.816ZM75.7733 99.196C74.5613 99.196 73.7453 98.272 73.7453 96.94C73.7453 95.608 74.5613 94.684 75.7733 94.684C76.9733 94.684 77.7893 95.62 77.7893 96.964C77.7893 98.272 76.9733 99.196 75.7733 99.196ZM79.8497 97.984C79.8737 99.328 80.9657 100.072 82.5497 100.072C83.9057 100.072 84.9977 99.364 84.9977 98.224C84.9977 96.868 83.8817 96.604 82.4897 96.436C81.6017 96.328 80.9537 96.244 80.9537 95.572C80.9537 95.008 81.5297 94.648 82.3817 94.648C83.2217 94.648 83.7977 95.116 83.8697 95.74H84.8777C84.7817 94.468 83.7737 93.82 82.3817 93.808C81.0017 93.796 79.9817 94.528 79.9817 95.644C79.9817 96.916 81.0737 97.192 82.4417 97.36C83.3537 97.468 84.0257 97.528 84.0257 98.296C84.0257 98.86 83.4137 99.232 82.5497 99.232C81.5297 99.232 80.8937 98.704 80.8577 97.984H79.8497ZM3.384 117.56C2.04 115.808 1.536 114.236 1.536 112.4C1.536 110.576 2.052 109.004 3.384 107.24H2.304C1.068 108.848 0.564 110.588 0.564 112.4C0.564 114.248 1.08 115.964 2.304 117.56H3.384ZM3.86166 116H9.64566V115.076H5.25366L7.77366 112.94C8.76966 112.088 9.58566 111.296 9.58566 110.036C9.58566 108.524 8.52966 107.528 6.71766 107.528C5.02566 107.528 3.90966 108.656 3.90966 110.24V110.456H4.91766V110.3C4.91766 109.148 5.63766 108.428 6.71766 108.428C7.83366 108.428 8.55366 109.028 8.55366 110.084C8.55366 110.996 7.90566 111.644 7.12566 112.316L3.86166 115.148V116ZM15.2362 116H16.3162L17.8522 111.452L19.3282 116H20.4082L22.4482 109.88H21.4522L19.8562 114.86L18.3322 109.88H17.4202L15.8122 114.86L14.2882 109.88H13.2562L15.2362 116ZM25.8397 116.072C27.2557 116.072 28.2637 115.172 28.5037 113.948H27.5077C27.3037 114.74 26.6317 115.196 25.7677 115.196C24.6037 115.196 23.8717 114.38 23.8357 113.216V113.132H28.5877C28.5997 112.988 28.6117 112.844 28.6117 112.712C28.5637 110.9 27.4357 109.808 25.7077 109.808C23.9797 109.808 22.8037 111.08 22.8037 112.94C22.8037 114.812 24.0037 116.072 25.8397 116.072ZM23.8957 112.316C23.9917 111.32 24.8197 110.672 25.7197 110.672C26.7397 110.672 27.4597 111.26 27.5557 112.316H23.8957ZM32.5545 116.072C33.9705 116.072 34.9785 115.172 35.2185 113.948H34.2225C34.0185 114.74 33.3465 115.196 32.4825 115.196C31.3185 115.196 30.5865 114.38 30.5505 113.216V113.132H35.3025C35.3145 112.988 35.3265 112.844 35.3265 112.712C35.2785 110.9 34.1505 109.808 32.4225 109.808C30.6945 109.808 29.5185 111.08 29.5185 112.94C29.5185 114.812 30.7185 116.072 32.5545 116.072ZM30.6105 112.316C30.7065 111.32 31.5345 110.672 32.4345 110.672C33.4545 110.672 34.1745 111.26 34.2705 112.316H30.6105ZM36.4493 116H37.4333V114.392L38.6693 113.12L40.6973 116H41.8853L39.3293 112.412L41.7173 109.88H40.5053L37.4333 113.132V107.6H36.4493V116ZM42.1388 113.984C42.1628 115.328 43.2548 116.072 44.8388 116.072C46.1948 116.072 47.2868 115.364 47.2868 114.224C47.2868 112.868 46.1708 112.604 44.7788 112.436C43.8908 112.328 43.2428 112.244 43.2428 111.572C43.2428 111.008 43.8188 110.648 44.6708 110.648C45.5108 110.648 46.0868 111.116 46.1588 111.74H47.1668C47.0708 110.468 46.0628 109.82 44.6708 109.808C43.2908 109.796 42.2708 110.528 42.2708 111.644C42.2708 112.916 43.3628 113.192 44.7308 113.36C45.6428 113.468 46.3148 113.528 46.3148 114.296C46.3148 114.86 45.7028 115.232 44.8388 115.232C43.8188 115.232 43.1828 114.704 43.1468 113.984H42.1388ZM47.923 107.24C49.267 108.992 49.771 110.564 49.771 112.4C49.771 114.224 49.255 115.796 47.923 117.56H49.003C50.239 115.952 50.743 114.212 50.743 112.4C50.743 110.552 50.227 108.836 49.003 107.24H47.923Z'
-				}
-				fill={'currentcolor'}
-			/>
-			<path
-				d={
-					'M4.428 188.072C6.324 188.072 7.716 186.992 8.028 185.252H6.924C6.612 186.428 5.712 187.112 4.416 187.112C2.652 187.112 1.572 185.732 1.572 183.8C1.572 181.856 2.7 180.488 4.464 180.488C5.724 180.488 6.612 181.184 6.936 182.348H8.04C7.728 180.608 6.372 179.528 4.512 179.528C2.076 179.528 0.504 181.256 0.504 183.8C0.504 186.368 2.016 188.072 4.428 188.072ZM13.4287 184.94C13.4287 186.32 12.7567 187.172 11.6407 187.172C10.6207 187.172 10.0447 186.5 10.0447 185.228V181.88H9.06066V185.288C9.06066 187.316 10.1287 188.072 11.4727 188.072C12.4807 188.072 13.0687 187.64 13.4287 187.184L13.5607 188H14.4127V181.88H13.4287V184.94ZM18.586 181.88C17.494 181.88 17.026 182.336 16.738 182.804L16.606 181.88H15.754V188H16.738V184.88C16.738 183.908 17.086 182.84 18.37 182.84H18.862V181.88H18.586ZM21.5122 188H22.7002L24.9922 181.88H23.9602L22.1242 186.884L20.2762 181.88H19.2202L21.5122 188ZM28.2772 188.072C29.6932 188.072 30.7012 187.172 30.9412 185.948H29.9452C29.7412 186.74 29.0692 187.196 28.2052 187.196C27.0412 187.196 26.3092 186.38 26.2732 185.216V185.132H31.0252C31.0372 184.988 31.0492 184.844 31.0492 184.712C31.0012 182.9 29.8732 181.808 28.1452 181.808C26.4172 181.808 25.2412 183.08 25.2412 184.94C25.2412 186.812 26.4412 188.072 28.2772 188.072ZM26.3332 184.316C26.4292 183.32 27.2572 182.672 28.1572 182.672C29.1772 182.672 29.8972 183.26 29.9932 184.316H26.3332ZM37.7726 188H39.0326L42.0566 179.6H40.9766L38.4206 186.764L35.8526 179.6H34.7486L37.7726 188ZM45.0115 188.072C46.7995 188.072 48.0475 186.8 48.0475 184.94C48.0475 183.08 46.7995 181.808 45.0115 181.808C43.2235 181.808 41.9755 183.08 41.9755 184.94C41.9755 186.8 43.2235 188.072 45.0115 188.072ZM45.0115 187.196C43.7995 187.196 42.9835 186.248 42.9835 184.94C42.9835 183.632 43.7995 182.684 45.0115 182.684C46.2235 182.684 47.0395 183.632 47.0395 184.94C47.0395 186.248 46.2235 187.196 45.0115 187.196ZM49.5616 186.404C49.5616 187.484 49.9816 188 51.1576 188H52.2256V187.1H51.2536C50.7256 187.1 50.5456 186.908 50.5456 186.392V182.78H52.1296V181.88H50.5456V180.164H49.5616V181.88H48.4576V182.78H49.5616V186.404ZM55.7459 188.072C57.1619 188.072 58.1699 187.172 58.4099 185.948H57.4139C57.2099 186.74 56.5379 187.196 55.6739 187.196C54.5099 187.196 53.7779 186.38 53.7419 185.216V185.132H58.4939C58.5059 184.988 58.5179 184.844 58.5179 184.712C58.4699 182.9 57.3419 181.808 55.6139 181.808C53.8859 181.808 52.7099 183.08 52.7099 184.94C52.7099 186.812 53.9099 188.072 55.7459 188.072ZM53.8019 184.316C53.8979 183.32 54.7259 182.672 55.6259 182.672C56.6459 182.672 57.3659 183.26 57.4619 184.316H53.8019ZM64.7494 188H65.9974L67.9054 181.34L69.8254 188H71.0734L73.5334 179.6H72.4534L70.4734 186.656L68.4814 179.6H67.4254L65.3614 186.656L63.3934 179.6H62.3014L64.7494 188ZM76.6873 188.072C78.1033 188.072 79.1113 187.172 79.3513 185.948H78.3553C78.1513 186.74 77.4793 187.196 76.6153 187.196C75.4513 187.196 74.7193 186.38 74.6833 185.216V185.132H79.4353C79.4473 184.988 79.4593 184.844 79.4593 184.712C79.4113 182.9 78.2833 181.808 76.5553 181.808C74.8273 181.808 73.6513 183.08 73.6513 184.94C73.6513 186.812 74.8513 188.072 76.6873 188.072ZM74.7433 184.316C74.8393 183.32 75.6673 182.672 76.5673 182.672C77.5873 182.672 78.3073 183.26 78.4033 184.316H74.7433ZM83.4022 188.072C84.8182 188.072 85.8262 187.172 86.0662 185.948H85.0702C84.8662 186.74 84.1942 187.196 83.3302 187.196C82.1662 187.196 81.4342 186.38 81.3982 185.216V185.132H86.1502C86.1622 184.988 86.1742 184.844 86.1742 184.712C86.1262 182.9 84.9982 181.808 83.2702 181.808C81.5422 181.808 80.3662 183.08 80.3662 184.94C80.3662 186.812 81.5662 188.072 83.4022 188.072ZM81.4582 184.316C81.5542 183.32 82.3822 182.672 83.2822 182.672C84.3022 182.672 85.0222 183.26 85.1182 184.316H81.4582ZM87.297 188H88.281V186.392L89.517 185.12L91.545 188H92.733L90.177 184.412L92.565 181.88H91.353L88.281 185.132V179.6H87.297V188ZM92.9864 185.984C93.0104 187.328 94.1024 188.072 95.6864 188.072C97.0424 188.072 98.1344 187.364 98.1344 186.224C98.1344 184.868 97.0184 184.604 95.6264 184.436C94.7384 184.328 94.0904 184.244 94.0904 183.572C94.0904 183.008 94.6664 182.648 95.5184 182.648C96.3584 182.648 96.9344 183.116 97.0064 183.74H98.0144C97.9184 182.468 96.9104 181.82 95.5184 181.808C94.1384 181.796 93.1184 182.528 93.1184 183.644C93.1184 184.916 94.2104 185.192 95.5784 185.36C96.4904 185.468 97.1624 185.528 97.1624 186.296C97.1624 186.86 96.5504 187.232 95.6864 187.232C94.6664 187.232 94.0304 186.704 93.9944 185.984H92.9864Z'
-				}
-				fill={'currentcolor'}
-			/>
-			<rect
-				x={'442'}
-				y={'56'}
-				width={'2'}
-				height={'176'}
-				fill={'#999'}
-			/>
-			<rect
-				x={'474'}
-				y={'24'}
-				width={'2'}
-				height={'240'}
-				fill={'#999'}
-			/>
-			<rect
-				x={'914'}
-				y={'24'}
-				width={'2'}
-				height={'240'}
-				fill={'#999'}
-			/>
-		</svg>
-	);
-}
