--- apps/ycrv/constants/menu.ts
@@ -1,4 +1,0 @@
-export const YCRV_MENU = [
-	{path: '/ycrv', label: 'yCRV'},
-	{path: '/ycrv/about', label: 'About'}
-];
