--- public/apps/ycrv-manifest.json
@@ -1,35 +1,0 @@
-{
-	"name": "yCRV",
-	"short_name": "yCRV",
-	"description": "The best crv yields in defi. *mic drop*",
-	"iconPath": "./favicons/favicon.svg",
-	"locale": "en-US",
-	"uri": "https://yearn.fi/ycrv/",
-	"og": "https://yearn.fi/apps/ycrv-og.png",
-	"twitter": "@iearnfinance",
-	"github": "https://github.com/Majorfi/y",
-	"icons": [
-		{
-			"src": "./favicons/android-icon-192x192.png",
-			"sizes": "192x192",
-			"type": "image/png",
-			"purpose": "any maskable"
-		},
-		{
-			"src": "./favicons/android-icon-144x144.png",
-			"sizes": "144x144",
-			"type": "image/png"
-		},
-		{
-			"src": "./favicons/android-icon-512x512.png",
-			"sizes": "512x512",
-			"type": "image/png"
-		}
-	],
-	"theme_color": "#000000",
-	"background_color": "#000000",
-	"title_color": "#ffffff",
-	"start_url": "/",
-	"display": "standalone",
-	"orientation": "portrait"
-}
